some modification

+++ database/migrations/2023_07_23_153445_create_factures_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->string('ref');
             $table->enum('type_fact', ['facture','bon','bon_livraison','bon_cmd','facture_d_avoir','facture_proforma','devis']);
             $table->date('date_facture')->nullable();
-            $table->float('ttc');
-            $table->float('ttva');
-            $table->float('tht');
+            $table->double('ttc', 15, 2);
+            $table->double('ttva', 15, 2);
+            $table->double('tht', 15, 2);
             $table->foreignId('client_id')->references('id')->on('clients')->onDelete('cascade');
             $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('exemple')->default('ex1');
